Refactor project folder and code for backend
