+++ backend/app/routes/query.py
@@ -1,0 +1,11 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+
+@router.post("/query")
+async def query_llm(question: str):
+    return {
+        "question": question,
+        "answer": "This is a placeholder response (LLM not connected yet)"
+    }
